separate ticket card in to different components
